+++ CHANGELOG.md
@@ -5,5 +5,5 @@ This repository contains multiple packages. Changelogs are maintained per packag
 - [`@explodingcamera/css`](packages/css/CHANGELOG.md)
 - [`minify-literals`](packages/minify-literals/CHANGELOG.md)
 - [`rollup-plugin-minify-template-literals`](packages/rollup-plugin-minify-template-literals/CHANGELOG.md)
-- [`simplejsx`](packages/simplejsx/CHANGELOG.md)
+- [`microjsx`](packages/microjsx/CHANGELOG.md)
 - [`spaify`](packages/spaify/CHANGELOG.md)